Implemented z-buffer

+++ geometry/vector3.go
@@ -74,8 +74,8 @@ func (v Vector3) toPoint() Point {
 	return Point{int(v.X), int(v.Y)}
 }
 
-func (v Vector3) Distance(v2 Vector3) float64 {
-	return math.Sqrt(math.Pow(v.X-v2.X, 2) + math.Pow(v.Y-v2.Y, 2) + math.Pow(v.Z-v2.Z, 2))
+func (v Vector3) Distance(v2 Vector3) float32 {
+	return float32(math.Sqrt(math.Pow(v.X-v2.X, 2) + math.Pow(v.Y-v2.Y, 2) + math.Pow(v.Z-v2.Z, 2)))
 }
 
 /*
